% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makegroups.R
\name{make_groups}
\alias{make_groups}
\title{Mytitle}
\usage{
make_groups(names)
}
\arguments{
\item{names}{}
}
\value{
names_coupled
}
\description{
Here is my description for this function. Here is a new sentence.
}
