% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myComplex.R
\docType{class}
\name{myComplex-class}
\alias{myComplex-class}
\alias{myComplex}
\title{myComplex objects}
\description{
An S4 class to represent complex numbers
}
\section{Slots}{

\describe{
\item{\code{a}}{real part}

\item{\code{b}}{imaginary part}
}}

