% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{myComplex2-methods}
\alias{myComplex2-methods}
\alias{show,myComplex2-method}
\alias{real,myComplex2-method}
\title{Methods for \code{\link{myComplex2}} class}
\usage{
\S4method{show}{myComplex2}(object)

\S4method{real}{myComplex2}(object)
}
\arguments{
\item{object}{An object of class myComplex2}
}
\description{
Methods for \code{\link{myComplex2}} class
}
