% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{real}
\alias{real}
\title{Real part}
\usage{
real(object)
}
\arguments{
\item{object}{An object of class myComplex or myComplex2}
}
\value{
The real part of the complex number
}
\description{
Access the real part of a complex number
}
