% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testone.R
\name{testone}
\alias{testone}
\title{Test Function 1}
\usage{
testone(x)
}
\arguments{
\item{x}{a number or vector}
}
\value{
a returns a number or vector
}
\description{
This test function does not have a space after "document"
}
\details{
the function returns the square of the input
}
\examples{
b <- 10
testone(b)

}
