% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestPackage2.R
\docType{package}
\name{TestPackage2}
\alias{TestPackage2}
\alias{TestPackage2-package}
\title{TestPackage2: A package to add complex numbers}
\description{
The TestPackage2 package provides two S4 classes to represent complex numbers
and a function to add them.
}
