% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testtwo.R
\name{testtwo}
\alias{testtwo}
\title{Test Function 2}
\usage{
testtwo(y)
}
\arguments{
\item{y}{a number or vector}
}
\value{
a returns a number or vector
}
\description{
This test function \emph{does} have a space after "document"
}
\details{
the function returns twice the input
}
\examples{
b <- 10
testtwo(b)

}
