% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{myComplex-methods}
\alias{myComplex-methods}
\alias{show,myComplex-method}
\title{Methods for \code{\link{myComplex}} class}
\usage{
\S4method{show}{myComplex}(object)
}
\description{
Methods for \code{\link{myComplex}} class
}
